Use `Rank` type for `nranks`

+++ cpp/src/communicator/mpi.cpp
@@ -95,7 +95,7 @@ MPI::MPI(MPI_Comm comm) : comm_{comm}, logger_{this} {
     RAPIDSMP_MPI(MPI_Comm_rank(comm_, &rank));
     RAPIDSMP_MPI(MPI_Comm_size(comm_, &nranks));
     rank_ = rank;
-    nranks_ = static_cast<std::int32_t>(nranks);
+    nranks_ = nranks;
     check_mpi_thread_support();
 }
